added VerifyUserInput.java
the main aim of the added VerifyUserInput class is to ensure only legal inputs are accepted.

eg:
"abcd" should not be allowed as an email or a password

email and password requirements:
email - need @ symbol
password - at least 6 characters and must have both numbers and letters.

This is a proof of concept for Delegation strategy, static methods and the good practice of verifying user inputs.